docshell/test/mochitest/test_bug529119-1.html

<!DOCTYPE HTML>
<html>
<head>
<title>Test bug 529119</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("untriaged");

var workingURL = "http://mochi.test:8888/tests/docshell/test/mochitest/bug529119-window.html";
var faultyURL = "http://www.some-nonexistent-domain-27489274c892748217cn2384.com/";

var w = null;
var phase = 0;
var gotWrongPageOnTryAgainClick = false;

function delay(msec) {
  return new Promise(resolve => setTimeout(resolve, msec));
}

async function pollForPage(win) {
  // Start with polling after a delay, we might mistakenly take the current page
  // as an expected one.
  await delay(1000);

  for (let i = 0; i < 200; i++) {
    await delay(100);

    try {
      win.document.title;
    } catch (ex) {
      return true;
    }
  }
  return false;
}

async function windowLoaded() {
  switch (phase) {
    case 0:
      /* 2. We have succeededfully loaded a page, now go to a faulty URL */
      window.setTimeout(function() {
        w.location.href = faultyURL;
      }, 0);

      phase = 1;

      ok(await pollForPage(w), "Waiting for error page succeeded");

      /* 3. now, while we are on the error page, try to reload it, actually
         click the "Try Again" button */
      SpecialPowers.spawn(w, [], () => this.content.location.reload());

      ok(await pollForPage(w), "Waiting for error page succeeded");

      /* 4-finish, check we are still on the error page */
      is(await SpecialPowers.spawn(w, [], () => this.content.location.href),
         faultyURL,
         "Is on an error page");
      is(gotWrongPageOnTryAgainClick, false,
        "Must not get www.example.com page on reload of an error page");
      w.close();
      SimpleTest.finish();
      break;

    case 1:
      /* 4-check, we must not get here! */
      gotWrongPageOnTryAgainClick = true;
      break;
  }
}

function startTest() {
  /* 1. load a URL that leads to an error page */
  w = window.open(workingURL);
}

</script>
</head>
<body onload="startTest();">
</body>
</html>
